made blog post title not unique and made updating of a blog post to be done by the user while deleting of blog post to be done by either the user or superuser resolves #50 #51

+++ bigfastapi/blog.py
@@ -14,26 +14,22 @@ from fastapi_pagination import Page, add_pagination, paginate
 
 app = APIRouter(tags=["Blog"])
 
-@app.post("/blog", response_model=schema.Blog)
-def create_blogpost(blog: schema.Blog, user: user_schema.User = fastapi.Depends(is_authenticated), db: orm.Session = fastapi.Depends(get_db)):
+@app.post("/blog", response_model=schema.BlogPost)
+def create_blogpost(blog: schema.BlogPost, user: user_schema.User = fastapi.Depends(is_authenticated), db: orm.Session = fastapi.Depends(get_db)):
     
     """Create a new blog
     
     Returns:
-        schema.Blog: Details of the newly created blog
+        schema.BlogPost: Details of the newly created blog
     """
-    
-    db_blog = model.get_blog_by_title(title=blog.title, db=db)
-    if db_blog:
-        raise fastapi.HTTPException(status_code=400, detail="Blog title already exists")
-    
-    blog = model.Blog(id=uuid4().hex, title=blog.title, content=blog.content, creator=user.id)
+   
+    blog = model.BlogPost(id=uuid4().hex, title=blog.title, content=blog.content, creator=user.id)
     db.add(blog)
     db.commit()
     db.refresh(blog)
-    return schema.Blog.from_orm(blog)
+    return schema.BlogPost.from_orm(blog)
 
-@app.get("/blog/{blog_id}", response_model=schema.Blog)
+@app.get("/blog/{blog_id}", response_model=schema.BlogPost)
 def get_blog(blog_id: str, db: orm.Session = fastapi.Depends(get_db)):
 
     """Get the details of a blog
@@ -42,24 +38,24 @@ def get_blog(blog_id: str, db: orm.Session = fastapi.Depends(get_db)):
         blog_id (str): the id of the blog
 
     Returns:
-        schema.Blog: Details of the requested blog
+        schema.BlogPost: Details of the requested blog
     """
 
-    return db.query(model.Blog).filter(model.Blog.id == blog_id).first()
+    return model.blogpost_selector(id=blog_id, db=db)
 
-@app.get("/blogs", response_model=Page[schema.Blog])
+@app.get("/blogs", response_model=Page[schema.BlogPost])
 def get_all_blogs(db: orm.Session = fastapi.Depends(get_db)):
 
     """Get all the blogs in the database
     
     Returns:
-        List of schema.Blog: A list of all the blogs
+        List of schema.BlogPost: A list of all the blogs
     """
 
-    blogs = db.query(model.Blog).all()
-    return paginate(list(map(schema.Blog.from_orm, blogs)))
+    blogs = db.query(model.BlogPost).all()
+    return paginate(list(map(schema.BlogPost.from_orm, blogs)))
 
-@app.get("/blogs/{user_id}", response_model=List[schema.Blog])
+@app.get("/blogs/{user_id}", response_model=List[schema.BlogPost])
 def get_user_blogposts(user_id: str, db: orm.Session = fastapi.Depends(get_db)):
 
     """Get all the blogs created by a user
@@ -68,51 +64,49 @@ def get_user_blogposts(user_id: str, db: orm.Session = fastapi.Depends(get_db)):
         user_id (str): the id of the user
 
     Returns:
-        List of schema.Blog: A list of all the blogs created by the user
+        List of schema.BlogPost: A list of all the blogs created by the user
     """
 
-    user_blogs = db.query(model.Blog).filter(model.Blog.creator == user_id).all()
-    return list(map(schema.Blog.from_orm, user_blogs))
+    user_blogs = db.query(model.BlogPost).filter(model.BlogPost.creator == user_id).all()
+    return list(map(schema.BlogPost.from_orm, user_blogs))
     
-@app.put("/blog/{blog_id}", response_model=schema.Blog)
-def update_blogpost(blog: schema.Blog, blog_id: str, user: user_schema.User = fastapi.Depends(is_authenticated), db: orm.Session = fastapi.Depends(get_db)):
+@app.put("/blog/{blog_id}", response_model=schema.BlogPost)
+def update_blogpost(blog: schema.BlogPost, blog_id: str, user: user_schema.User = fastapi.Depends(is_authenticated), db: orm.Session = fastapi.Depends(get_db)):
 
-    """Update the details of a blog
+    """Update the details of a blog by the user or superuser
     
     Args:
         blog_id (str): the id of the blog to be updated
 
     Returns:
-        schema.Blog: Refreshed data of the updated blog
+        schema.BlogPost: Refreshed data of the updated blog
     """
 
-    blog_db = model.blog_selector(user=user, id=blog_id, db=db)
+    blog_db = model.blogpost_selector(id=blog_id, db=db)
 
     if blog_db == None:
         raise fastapi.HTTPException(status_code=404, detail="Blog does not exist")
 
+    if blog_db.creator != user.id:
+        raise fastapi.HTTPException(status_code=403, detail="You are not the creator of the blog post")
+    
     if blog.content != "":
         blog_db.content = blog.content
 
     if blog.title != "":
-        blog_db_title = model.get_blog_by_title(title=blog.title, db=db)
-
-        if blog_db_title:
-            raise fastapi.HTTPException(status_code=400, detail="Blog title already in use")
-        else:
-            blog_db.title = blog.title
+        blog_db.title = blog.title
 
     blog_db.last_updated = datetime.utcnow()
 
     db.commit()
     db.refresh(blog_db)
 
-    return schema.Blog.from_orm(blog_db)
+    return schema.BlogPost.from_orm(blog_db)
 
 @app.delete("/blog/{blog_id}")
 def delete_blogpost(blog_id: str, user: user_schema.User = fastapi.Depends(is_authenticated), db: orm.Session = fastapi.Depends(get_db)):
     
-    """Delete a blog from the database
+    """Delete a blog from the database by the user or superuser
     
     Args:
         blog_id (str): the id of the blog to be deleted
@@ -121,13 +115,17 @@ def delete_blogpost(blog_id: str, user: user_schema.User = fastapi.Depends(is_au
         object (dict): successfully deleted
     """
     
-    blog = model.blog_selector(user=user, id=blog_id, db=db)
+    blog = model.blogpost_selector(id=blog_id, db=db)
     if blog == None:
         raise fastapi.HTTPException(status_code=404, detail="Blog does not exist")
-        
-    db.delete(blog)
-    db.commit()
+       
+    if blog.creator == user.id or user.is_superuser == True:
+        db.delete(blog)
+        db.commit()
 
-    return {"message":"successfully deleted"}
+        return {"message":"successfully deleted"}
+    else:
+        raise fastapi.HTTPException(status_code=403, detail="You are not the allowed to delete the blog post")
+
 
 add_pagination(app)
